refacto checkers

+++ src/install/biglybt/biglybt.bash
@@ -1,19 +1,23 @@
 ##!/usr/bin/env bash
 include utils/move_to_system.bash
 include utils/download.bash
+include utils/checker.bash
 
 install_biglybt() {
-  local tmp_biglybt_installer="/tmp/BiglyBT_Installer.sh"
-  download_biglybt_installer "$tmp_biglybt_installer"
-  install_expect
-  install_biglybt_from_installer_without_display "$tmp_biglybt_installer" "$APP_DIR/biglybt"
-  rm "$tmp_biglybt_installer"
-  move_from_home_to_system .biglybt
+    checker check_biglybt_install
+    if failure; then
+        local tmp_biglybt_installer="/tmp/BiglyBT_Installer.sh"
+        download_biglybt_installer "$tmp_biglybt_installer"
+        install_expect
+        install_biglybt_from_installer_without_display "$tmp_biglybt_installer" "$APP_DIR/biglybt"
+        rm "$tmp_biglybt_installer"
+    fi
+    move_from_home_to_system .biglybt
 }
 
 check_biglybt() {
-  [[ $(readlink -f "$HOME/.biglybt") = "$SYSTEM_DIR/biglybt" ]]
-  [[ -f "$APP_DIR/biglybt/biglybt" ]]
+    check_biglybt_install
+    check_biglybt_symlink
 }
 
 download_biglybt_installer() {
@@ -31,4 +35,12 @@ install_biglybt_from_installer_without_display() {
 install_expect() {
     sudo apt update
     sudo apt-get install -y expect
+}
+
+check_biglybt_install(){
+    [[ -f "$APP_DIR/biglybt/biglybt" ]]
+}
+
+check_biglybt_symlink(){
+    [[ $(readlink -f "$HOME/.biglybt") = "$SYSTEM_DIR/biglybt" ]]
 }
